+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 CHANGELOG
 =========
+0.1.0
+-----
+- fix bug with global logger
+- fix display of progress
+- new field "log_file_path
+- changed default timeout = 1
+
 0.0.1
 -----